Add PhasePortrait3DManager. Merge 2D and 3D phaseportrait managers into same file.
